moved cpu util API
