+++ src/variorum/config_architecture.c
@@ -356,7 +356,6 @@ void variorum_init_func_ptrs()
         g_platform[i].variorum_print_gpu_utilization = NULL;
         g_platform[i].variorum_monitoring = NULL;
         g_platform[i].variorum_get_node_power_json = NULL;
-        g_platform[i].variorum_get_node_util_json = NULL;
         g_platform[i].variorum_get_node_power_domain_info_json = NULL;
         g_platform[i].variorum_print_energy = NULL;
     }
