+++ src/variorum/variorum.c
@@ -10,7 +10,12 @@
 #include <config_architecture.h>
 #include <variorum.h>
 #include <variorum_error.h>
+#include <sys/time.h>
+#include <sys/resource.h>
 
+int state = 0;
+long prevTime = 0;
+long prevMem = 0;
 int g_socket;
 int g_core;
 
@@ -1010,7 +1015,7 @@ int variorum_get_node_power_json(char **get_power_obj_str)
     return err;
 }
 
-int variorum_get_node_util_json(char **get_util_obj_str)
+int variorum_get_node_utilization_json(char **get_util_obj_str)
 {
     int err = 0;
     int i;
@@ -1020,49 +1025,51 @@ int variorum_get_node_util_json(char **get_util_obj_str)
         return -1;
     }
 
-    // Obtain the index corresponding to the primary platform.
-    for (i = 0; i < P_NUM_PLATFORMS; i++)
-    {
-#ifdef VARIORUM_WITH_INTEL_CPU
-        i = P_INTEL_CPU_IDX;
-        break;
-#endif
-#ifdef VARIORUM_WITH_IBM_CPU
-        i = P_IBM_CPU_IDX;
-        break;
-#endif
-#ifdef VARIORUM_WITH_AMD_CPU
-        i = P_AMD_CPU_IDX;
-        break;
-#endif
-#ifdef VARIORUM_WITH_ARM_CPU
-        i = P_ARM_CPU_IDX;
-        break;
-#endif
-    }
-
-    if (g_platform[i].variorum_get_node_util_json == NULL)
-    {
-        variorum_error_handler("Feature not yet implemented or is not supported",
-                               VARIORUM_ERROR_FEATURE_NOT_IMPLEMENTED,
-                               getenv("HOSTNAME"), __FILE__,
-                               __FUNCTION__, __LINE__);
-        // For the JSON functions, we return a -1 here, so users don't need
-        // to explicitly check for NULL strings.
-        return -1;
-    }
-
-    err = g_platform[i].variorum_get_node_util_json(get_util_obj_str);
-    if (err)
-    {
-        return -1;
-    }
     err = variorum_exit(__FILE__, __FUNCTION__, __LINE__);
     if (err)
     {
         return -1;
     }
-    return err;
+    int ru;
+    char hostname[1024];
+    struct timeval tv, sstart, send, ustart, rutime;
+    struct rusage rusge;
+    uint64_t ts;
+    long mem, utime;
+    json_t *get_util_obj = json_object();
+
+    gethostname(hostname, 1024);
+    gettimeofday(&tv, NULL);
+    ts = tv.tv_sec * (uint64_t)1000000 + tv.tv_usec;
+    json_object_set_new(get_util_obj, "host", json_string(hostname));
+    json_object_set_new(get_util_obj, "timestamp", json_integer(ts));
+
+    ru = getrusage(RUSAGE_SELF, &rusge);
+    if (ru < 0)
+    {
+        printf("Failed to get utilizations\n");
+        return -1;
+    }
+
+    rutime = rusge.ru_utime;
+    utime = ((rutime.tv_sec * 1000000) + rutime.tv_usec);
+    mem = rusge.ru_maxrss;
+
+    if (state)
+    {
+        utime -= prevTime;
+        mem -= prevMem;
+    }
+
+    prevTime = utime;
+    prevMem = mem;
+    json_object_set_new(get_util_obj, "memory util", json_integer(mem));
+    json_object_set_new(get_util_obj, "ru time", json_integer(utime));
+    *get_util_obj_str = json_dumps(get_util_obj, 0);
+    json_decref(get_util_obj);
+    state = 1;
+    return 0;
+
 }
 // The variorum_get_node_power_domain_info_json is a node-level API, and cannot
 // be implemented at a per-component (eg CPU, GPU) level. This can only be available
