+++ src/variorum/variorum.h
@@ -7,6 +7,7 @@
 #define VARIORUM_H_INCLUDE
 
 #include <stdio.h>
+#include <jansson.h>
 
 /// @brief Collect power limits and energy usage for both the package and DRAM
 /// domains.
@@ -545,8 +546,7 @@ int variorum_get_node_power_json(char **get_power_obj_str);
 ///
 /// @param [out] output String (passed by refrence) that contains node-level utilization information.
 /// returns 0 if successful, otherwise -1.
-int variorum_get_node_util_json(char **get_util_obj_str);
-
+int variorum_get_node_utilization_json(char **get_util_obj_str);
 /// @brief Populate a string in JSON format with measurable and controllable
 /// power domains, along with the ranges.
 ///
